# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorDetail": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				]
			},
			"ErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"$ref": "#/components/schemas/ErrorDetail"
					},
					"timestamp": {
						"type": "string"
					},
					"developer": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NullSuccessResponse": {
				"$ref": "#/components/schemas/SuccessResponse_null_"
			},
...
			"LocationType": {
				"type": "string",
				"enum": [
					"customer",
					"delivery_partner"
				]
			},
			"ILocationPoint": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"Point"
						],
						"nullable": false
					},
					"coordinates": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"type",
					"coordinates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrackingStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive",
					"completed"
				]
			},
			"ILocationHistory": {
				"properties": {
					"coordinates": {
						"$ref": "#/components/schemas/ILocationPoint"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"accuracy": {
						"type": "number",
						"format": "double"
					},
					"heading": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"coordinates",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILiveLocation": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/LocationType"
					},
					"orderId": {
						"type": "string"
					},
					"currentLocation": {
						"$ref": "#/components/schemas/ILocationPoint"
					},
					"lastUpdated": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/TrackingStatus"
					},
					"sessionId": {
						"type": "string"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"accuracy": {
						"type": "number",
						"format": "double"
					},
					"heading": {
						"type": "number",
						"format": "double"
					},
					"batteryLevel": {
						"type": "number",
						"format": "double"
					},
					"locationHistory": {
						"items": {
							"$ref": "#/components/schemas/ILocationHistory"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"userType",
					"currentLocation",
					"lastUpdated",
					"status",
					"sessionId",
					"locationHistory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_ILiveLocation_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/ILiveLocation"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStartTrackingRequest": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/LocationType"
					},
					"orderId": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"sessionId": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userType",
					"latitude",
					"longitude",
					"sessionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateLocationRequest": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/LocationType"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"sessionId": {
						"type": "string"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"accuracy": {
						"type": "number",
						"format": "double"
					},
					"heading": {
						"type": "number",
						"format": "double"
					},
					"batteryLevel": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"userType",
					"latitude",
					"longitude",
					"sessionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_ILiveLocation-or-null_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ILiveLocation"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_ILiveLocation-Array_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/ILiveLocation"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ICoupon": {
				"properties": {
					"code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"discountType": {
						"type": "string",
						"enum": [
							"percentage",
							"fixed"
						]
					},
					"discountValue": {
						"type": "number",
						"format": "double"
					},
					"maxDiscountAmount": {
						"type": "number",
						"format": "double"
					},
					"minOrderValue": {
						"type": "number",
						"format": "double"
					},
					"validFrom": {
						"type": "string",
						"format": "date-time"
					},
					"validUntil": {
						"type": "string",
						"format": "date-time"
					},
					"totalUsageLimit": {
						"type": "number",
						"format": "double"
					},
					"timesUsed": {
						"type": "number",
						"format": "double"
					},
					"usageLimitPerUser": {
						"type": "number",
						"format": "double"
					},
					"usedBy": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"code",
					"description",
					"discountType",
					"discountValue",
					"minOrderValue",
					"validFrom",
					"validUntil",
					"totalUsageLimit",
					"timesUsed",
					"usageLimitPerUser",
					"usedBy",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaginated": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalRecord": {
						"type": "number",
						"format": "double"
					},
					"totalPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"page",
					"limit",
					"totalRecord",
					"totalPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessDataResponse_ICoupon_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/ICoupon"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ICoupon_": {
				"properties": {
					"code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"discountType": {
						"type": "string",
						"enum": [
							"percentage",
							"fixed"
						]
					},
					"discountValue": {
						"type": "number",
						"format": "double"
					},
					"maxDiscountAmount": {
						"type": "number",
						"format": "double"
					},
					"minOrderValue": {
						"type": "number",
						"format": "double"
					},
					"validFrom": {
						"type": "string",
						"format": "date-time"
					},
					"validUntil": {
						"type": "string",
						"format": "date-time"
					},
					"totalUsageLimit": {
						"type": "number",
						"format": "double"
					},
					"timesUsed": {
						"type": "number",
						"format": "double"
					},
					"usageLimitPerUser": {
						"type": "number",
						"format": "double"
					},
					"usedBy": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"SuccessDataResponse__message-string__": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Enter JWT token with `Bearer <token>`"
			}
		}
	},
	"info": {
		"title": "milqit-ecom-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/location/start": {
			"post": {
				"operationId": "StartTracking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_ILiveLocation_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Start tracking session (REST API fallback)",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStartTrackingRequest"
							}
						}
					}
				}
			}
		},
		"/location/update": {
			"post": {
				"operationId": "UpdateLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_ILiveLocation_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update location (REST API fallback)",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateLocationRequest"
							}
						}
					}
				}
			}
		},
		"/location/current/{userId}": {
			"get": {
				"operationId": "GetCurrentLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_ILiveLocation-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get current location",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "orderId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/location/stop": {
			"post": {
				"operationId": "StopTracking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NullSuccessResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Stop tracking",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"orderId": {
										"type": "string"
									},
									"sessionId": {
										"type": "string"
									},
									"userId": {
										"type": "string"
									}
								},
								"required": [
									"sessionId",
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/location/nearby": {
			"get": {
				"operationId": "FindNearbyDeliveryPartners",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_ILiveLocation-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Find nearby delivery partners",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "latitude",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "longitude",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDistance",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/location/history/{userId}": {
			"get": {
				"operationId": "GetLocationHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_ILiveLocation-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get location history",
				"tags": [
					"Location Tracking"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "orderId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/coupons": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessDataResponse_ICoupon_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Coupons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICoupon"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pagination": {
											"$ref": "#/components/schemas/IPaginated"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/ICoupon"
											},
											"type": "array"
										}
									},
									"required": [
										"pagination",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Coupons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/admin/coupons/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessDataResponse_ICoupon_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Coupons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ICoupon_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessDataResponse__message-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Coupons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:5001/v1"
		}
	]
}